Splitting grib.py in gribfile.py and gribdata.py
